Add files via upload
Profile retrieving, editing, and creating is now functional. Also changed battery consumption lol

+++ RadotechSim/userprofile.cpp
@@ -1,6 +1,6 @@
 #include "userprofile.h"
 
-UserProfile::UserProfile(int id, QString n, int a, int h, int w) {
+UserProfile::UserProfile(int id, QString n, int a, double h, double w) {
 	this->id = id;
 	this->name = n;
 	this->age = a;
@@ -20,11 +20,11 @@ void UserProfile::deleteRecord(int idx) {
 int UserProfile::getId() { return id; }
 QString UserProfile::getName() { return name; }
 int UserProfile::getAge() { return age; }
-int UserProfile::getHeight() { return height; }
-int UserProfile::getWeight() { return weight; }
+double UserProfile::getHeight() { return height; }
+double UserProfile::getWeight() { return weight; }
 QVector<Record *>* UserProfile::getRecords() { return &records; }
 
 void UserProfile::setName(QString newName) { name = newName; }
 void UserProfile::setAge(int newAge) { age = newAge; }
-void UserProfile::setHeight(int newHeight) { height = newHeight; }
-void UserProfile::setWeight(int newWeight) { weight = newWeight; }
+void UserProfile::setHeight(double newHeight) { height = newHeight; }
+void UserProfile::setWeight(double newWeight) { weight = newWeight; }
